ファイル切り分け，体裁の整理
