+++ app/ws_connect.go
@@ -1,0 +1,58 @@
+// websocket通信関係の関数，構造体を記述する
+
+package main
+
+import (
+	"fmt"
+	"net/http"
+	"sync"
+
+	"github.com/gin-gonic/gin"
+	"github.com/gorilla/websocket"
+	"gorm.io/gorm"
+)
+
+// wsの保存用
+type WsMap struct {
+	sync.RWMutex
+	m map[*websocket.Conn]struct{}
+}
+
+// upgraderはHTTPをWSにするときに呼ばれる
+// ここで許可するoriginや接続時間を設定する
+var upgrader = websocket.Upgrader{
+	// TODO: ここ絶対直すこと
+	CheckOrigin: func(r *http.Request) bool {
+		return true
+	},
+}
+
+// ws用のエンドポイントのハンドラ
+func wshandler(db *gorm.DB, wsMap *WsMap, broadcastChan *BroadChan) func(*gin.Context) {
+	return func(c *gin.Context) {
+		// websocketで接続
+		ws, err := upgrader.Upgrade(c.Writer, c.Request, nil)
+		if err != nil {
+			fmt.Println(err)
+			return
+		}
+
+		// r.GETが終わった時にクローズ（クローズハンドラがないから特に何もしない）
+		defer ws.Close()
+
+		// r.GETが終わった時にwsをwsMapから削除
+		defer func(wsMap *WsMap) {
+			wsMap.Lock()
+			delete(wsMap.m, ws)
+			wsMap.Unlock()
+		}(wsMap)
+
+		// ブロードキャスト用にソケットを保存
+		wsMap.Lock()
+		wsMap.m[ws] = struct{}{}
+		wsMap.Unlock()
+
+		// クライアントからのwebsocketを処理
+		procClient(c, db, ws, broadcastChan)
+	}
+}
